refactor: move asset, robotics and 3D loaders
