+++ tests/test_integration/test_loaders/test_resource_container_loaders.py
@@ -16,7 +16,7 @@ from cognite.client.exceptions import CogniteAPIError
 
 from cognite_toolkit._cdf_tk.client import ToolkitClient
 from cognite_toolkit._cdf_tk.loaders import ContainerLoader, TimeSeriesLoader
-from cognite_toolkit._cdf_tk.prototypes.resource_loaders import ThreeDModelLoader
+from cognite_toolkit._cdf_tk.loaders._resource_loaders.three_d_model_loaders import ThreeDModelLoader
 from tests.test_integration.constants import RUN_UNIQUE_ID
